Sistema Porta/requirements.txt
djangorestframework==3.9.1
django
paho.mqtt
